On a snapp upgrade, we need to make the current version not-active and run all the hooks to make it no longer active in systemd, not startable, etc *before* we copy the data. This ensures that nothing writes data while we copy the data which may lead to data corruption.
This is not very atomic, but the design to have a data copy on upgrade
means we must have consistent data. Plus the data will be in a new
place so the apps needs to restart to look at this new place.

Without the data copy and per version data-dirs we could simply keep
the app running.