Move dependency handing earlier
New developers, such as myself, will find that go build doesn't really
work out of the box. The needed godeps steps is documented but it is
mentioned later in the README file. Inpatient developers may not read
everything top-to-bottom to see this patch moves the relevant section
into the correct spot.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>