Add addtrap() helper to run-checks
We use a trap to display "endmsg" which tells about success or
failure of the tests. However the script also uses a second trap
to do cleanup which overrides the first trap. This means no success
or failure message at the end. This branch fixes this by allowing
stacking of trap functions.