Add the concept of capability attributes.
Attributes are the primordial form of "properties" and "specifications"
that were discussed informally earlier. Since we're doing baby steps and
trying things out little by little, I want to introduce the concept of
capability attributes.

Attributes are intrinsically bound to types. So far types don't have a
representation but this if fine at this stage. Attributes will help us
make the first effective use of capabilities. My intent is to use them
as bits of data that can be used by applications to convey the location
of the resource (e.g. the path of the device to open: /dev/ttyUSB0).
Later on, they will also play the central role with enabling the
security parts that go along with this to work.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>