use EvalSymlinks to resolve 'current' to the correct version
'snappy policygen' takes a path to a yaml file as an argument. If the user
specifies /apps/foo/current/meta/package.yaml then the version is incorrectly
calculated to be 'current'. This issue was introduced in 8b04bd2d when fixing
CompareGeneratePolicyFromFile for sideloaded pkgs.

The fix adjusts parsePackageYamlFileWithVersion() to simply call EvalSymlinks()
to resolve the current symlink. Tested with store and sideloaded apps.