Introduce iotaType capability type for testing
Testing the capability repository requires operating with valid types
and capabilities. Originally the FileType was used for that but it was
never really designed as a proper type. To differentiate production
types (that we'll have later) and the testing type, I've introduced
iotaType and added it to the testing iotaRepo.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>