Add to-from JSON converters for caps.Capability
This patch adds two things.

First is the new type caps.CapabilityRepr which represents a capability
in a manner that is friendly to JSON serialization. In particular the
caps.Capability.Type field is replaced with TypeName (string). Struct
Tags are used to make JSON representation of CapabilityRepr look nice
and friendly.

The second thing is a pair of converter functions that assist in making
the change from one to the other.

Both are essential to the upcoming REST API.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>