Refactor error handling of client.Capabilities()
There are two changes here. First of all the flow is different, main
body of the code uses a switch statement to handle different response
types.  The error type is now tweaked to use errorResult to handle
errors in a way that preserves the error message produced by the server.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>