classic/run.go: do not kill all processes when exiting classic
As discussed on the sprint the classic mode will not kill all
processes that got started in the "snappy shell classic". Only
processes in the current terminal control group get killed.

We still run everything in a cgroup that is easy to identify
in order to allow the user to easily stop all processes via
"systemctl stop" though.