Update daemon/ package for no-more dirs.SnapIconsDir
The daemon package has two ways to get icons:
 /1.0/icons/[name]/icon
 /1.0/icons/[icon]

Now that we no longer download the icon from the store
we can name the icon for the package in a canonical way.
So the second entry point becomes just a alias (for
compatiblity) for the first one and the name that is
returned by the list command for the icon key in the json
is always the canonical package name.