Use "mask/unmask" when enabling/disabling the snappy-autopilot.timer
The snappy-autopilot.timer file is available on the system-[ab]
partition and also on the writable partition. This mean that when
running systemctl disable the symlink for the snappy-autopilot.timer
gets removed from the writable space. However on the next boot
the symlink is re-created from the writable-path logic that detects
that the OS has a file that is missing in the writable space.

By using mask/unmask this is fixed because with these commands
the files are kept on writable and they are linked to dev/null.