Re-add some support for /oem and the OEM snap type
We recently moved from OEM to gadget everywhere. This is great,
however some users will still have the oem snap installed and our
tests are currently relying on those too. This branch adds enough
support for the oem snap to support this use-case.

Note that once we move to the all-snap world fully we do not need
this branch anymore. The all-snap gadget snaps are already type
gadget and users will have to start with a fresh image anyway
here. In this light, *maybe* we should just ignore this commit
and wait until all-snap is real.