Pass the parent environemnt when re-execing for "sudo"
The "snappy shell classic" command checks if it is running as
a normal user and if that is the case it will re-exec itself
with the "sudo" command. Its using syscall.Exec() for this.

In order to have a working environment (so that e.g. TERM is
correct) we now pass the parent environment.