Start porting to the all-snap architecture
Version numbers in the all-snap kernel/os world are not longer
just integers but strings. Code updated to reflect that.

The base image needs to be created differently too, code updated
to reflect this. Unfortunately the u-d-f from
  lp:~snappy-dev/goget-ubuntu-touch/all-snaps
is not merged in 16.04 so for now it needs to be build manually.

To test the easiest option is to build an image once (e.g. via
the --output-dir option in the tests runner) and then run the
image with:

  $ qemu-system-x86_64 -enable-kvm -m 512 -nographic -net user -net nic,model=virtio -snapshot -snapshot -drive file=snappy-rolling-edge-latest.img,if=virtio -redir tcp:11022::22 -monitor none -serial stdio

and run the tests with:

  $ go run  integration-tests/main.go  --ip localhost -port 11022