Fix the list output after a "snappy update"
After running "snappy update" with an all-snap system the output
is often incorrect. The reboot notification is not displayed, the
date of the install is wrong etc.

The reason for this is that the Update{,All}() calls return a list
of applied updates that is then displayed. However the list of
applied updates is for the remote snaps. We need to convert the
remote snaps to the locally installed ones instead to get the
correct date and the correct reboot-required checks.