Do not run Verify() twice on a given package
The current code runs Verify() of the package in NewSnapFile()
and again in SnapFile.Install(). The second Verify() is not
needed because at this point it is either verified or the user
has overriden the verification check. We need to do the check
earlier in NewSnapFile to ensure that we do not touch unsigned
data that is potentially dangerous.