Add capability attribute types
This patch lays the work for replacing the per-capability-type list of
required attributes with a map of attribute types.

Each attribute type can perform a single operation, CheckValue() which can
validate and, if required, transform the value.

This change allows capabilities to have attributes with non-trivial
semantics, like a path attribute that dereferences symbolic links,
for example. This was found required by the earlier work on the bool-file
capability type.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>