Change the update test to use the new fake store
In the new all-snaps world we can no longer fake an update by
changing the installed version. So we generate a new version
of the installed ubuntu-core.canonical package instead.

Note that this needs enough space in /tmp so that the new snap
can be build. I ran it successfully with:

  $ qemu-system-x86_64 -enable-kvm -m 1500 -nographic -net user -net nic,model=virtio -snapshot -snapshot -drive file=snappy-rolling-edge-latest.img,if=virtio -redir tcp:11022::22