Rewrite capabilities to use interfaces
This large change discards the declarative capability types.
A caps.Capability is now an interface type
A caps.Type is now an interface type.

Concrete capabilities are separate go types (structs).  Capability type
knows how to make instances of that type using the caps.Type.Make()
function. To aid in serialization each Capability can now marshal itself
in a way that can be unmarshaled as CapabilityInfo. CapabilityInfo is
generic and simply expresses the public information about any
capability. The caps.Repository.MakeCapFromInfo() function can be used
to make a proper (typed) category.

The advantage of this model is that capabilities are self-contained
types with arbitrary code (not declarations) where appropriate.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>