Replace caps.Type struct with an interface
The capability type is now an interface. This allows each type to freely
define appropriate logic in a non-declarative way. Some small changes
related to that are the change of the Validate() method to Sanitize()
which can also explicitly mutate a capability (currently this is not
used yet).

Along with this change, caps.Capability.Type is replaced with .TypeName
which is simply a string so that capabilities can be serialized easily.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>