Check for valid path of bool-file capability type
To enhance the security of the bool-file capability it must not be
allowed to create an arbitrary bool-file (e.g. one pointing to /dev/sda
or /dev/mem). This is achieved with simple validation against a set of
regular expressions describing allowed values.

Currently two kinds of bool file can be created. They can point to:
- The value of a sysfs GPIO
- The brightness of a sysfs LED

Additional values can be added as we discover their utility.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>