Allow types to hand out security snippets
This patch commences the work on the security side of capabilities. Each
capability type now has a way to hand out "snippets" of security
information applicable to a given security system. The information
describes alterations of the security system needed to consume a given
capability.

The patch defines three security system names:
 - apparmor
 - seccomp
 - dbus

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>