Identify capabilities with pairs (snap, cap)
This patch makes capability name non-unique. Instead the pair (snap
name, capability name) is now the unique way to refer to capabilities in
the system.

This change is motivated by the need to resolve conflicts when more than
one snap wants to create a capability with a given name. Using the pair
makes the problem go away.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>