Model granted capabilities
Granted capabilities are modeled by the capability repository. The
repository stores a two-level map[CapabilityID]map[CapabilityID]bool.
The final value is dummy so the whole type acts as a set of pairs
(provided, consumed) where both fields represent the CapabilityID of the
provided and consumed capability.

This arrangement allows for multiple consumers of any capability and has
efficient lookup mechanism.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>