Sync the active versions of the OS and Kernel on boot
The fallover code may need to fallback on to the previous
OS/Kernel on boot. When this happens we also need to update
the "current" active version of the snap on disk. I.e. when
kernel-v2 fails to boot we want to revert the active kernel
in snappy list to the previous version.