daemon: add REST APIs for skills
REST APIs for skills are very simple and are comprised of only one
endpoint and two HTTP methods:

GET /2.0/skills

This returns a list of skills with basic information about each skill
and the users (slots) of that skill. Note that not everything is
returned. Some implementation details (apps and attributes) are hidden.

Example response:

    [
        {
            snap:  "canonical-pi2",
            type:  "bool-file”,
            name:  "pin-13",
            label: "Pin 13",
            slots: [
                {"snap": "keyboard-lights", "name": "capslock-led"}
            ]
        }
    ]

POST /2.0/skills

This provides an RPC-like interface for interacting with the skill system.
Available actions include "grant" and "revoke". Additional actions, included
only for development, include "{add,remove}-{skill,slot}".

Example request:

    {
        “action”: “grant” | “revoke”,
        “skill”:  {“snap”: “canonical-pi2”,   “name”: “pin-13”},
        “slot”:   {“snap”: “keyboard-lights”, “name”: “capslock-led”},
    }

When a daemon flag is set, the four additional actions are available:

 - add-skill
 - remove-skill
 - add-slot
 - remove-slot

Unlike the "grant" and "revoke" action, those actions use extended
syntax of the request, where the appropriate object (skill or slot) has
complete representation. This means that in addition to snap and name,
it must contain, type, label, attrs and apps.

For example:

    {
        “action”: “add-slot”,
        “slot”:   {
            “snap”: “keyboard-lights”,
            “name”: “capslock-led”,
            "type": "bool-file",
            "label": "Caps Lock LED",
            "attrs": {
                ...
            },
            "apps": "app1"
        },
    }

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>