cmd: construct fresh parser for each test
Since command objects store state inside and there is no way to reset
that state for each test the only alternative is to reconstruct a fresh
parser object for each test.

All of the command-specific init() functions are replaced by a Parser()
function that constructs the parser and adds each command.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>