testutil: handle interface types in Contains
This patch fixes Contains (and DeepContains) to handle collections of
interface types. In the past, the code would reject such checks as it
would look at the actual underlying type. Now, if the collection element
is an interface the test checks if the element being looked for
implements that interface.

The change is accompanied with a few simple tests to ensure both failing
and successful cases are working correctly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>