skills/types: add a way to find all the types
This patch adds a replacement to skills.LoadBuiltInTypes() that lives in
the skills/types package. The said package is going to contain all the
implemented skill types so it makes sense for the knowledge about
available types to be stored there to avoid circular imports back to the
skills/ package.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>