integration-tests: always use the built snapd when compiling binaries from branch
The commands for shutting down the system's snapd and setting up the compiled one
are moved from the snapdSuite SetUpTest method to the init() function of the tests
package (this way it gets executed once per boot) and it's only set up if we are
generating the binaries from the current branch.

The previous call to kill the process has been removed, we could keep it in the Test
function after the runner.TestingT(t, output) call, but in fact it wasn't stopping
the daemon forked by systemd-activate, the process being killed was the parent. This
will be handled in an upcoming branch.