client,cmd/snap: rename skills to interfaces
This patch changes the client-side interface to the interface system
(nee capabilities, skills) to use the new set of concepts (interfaces,
plugs, slots, connections and disconnections).

Internally, the wire format is identical to the old skill format so the
server doesn't have to change.

This patch also does trivial changes to the command line tools to use
the new function, type and field names.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>