Remove model's authority-id == brand-id restriction.
WARNING: probably a temporary change

Removing the restriction means that a trusted authority, e.g. Canonical,
can sign a brand's model assertion. (Devices must still verify that the
`model` assertion is signed by a trusted authority, as they must do if
it signed by the brand's key.)

It's quite likely this will be reverted in favour of a formal "signed on
behalf of" solution.

Also note that the model shouldn't actually have an `authority-id`
header field anyway.