client,daemon,docs: use plug/slot instead of name on the wire
The wire format used to refer to "name" (being either skill or slot name
in the past). Now the format is explicit and uses "plug" to refer to
"plug name" and "slot" to refer to "slot name".

Note that Go interface is left unchanged, it is addressed in a
subsequent branch as it has wider impact.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>