interfaces: allow security helpers to run commands
This patch adds a new method to the private securityHelper interface,
for returning a list of commands to run to make the security changes
"effective".

The only benefactor of this API is apparmor which in practice operates
on in-kernel profiles that have to be compiled and loaded with userspace
support.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>