interfaces: add SecurityCommandsForSnap()
This patch adds a new repository method for obtaining commands necessary
to "make effective" all the security changes. Currently the command
doesn't work as it depends on a way to enumerate applications within a
snap and this is not contained in the interface repository. This
information is likely be supplied by a cross-manager interaction.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>