client,cmd/snap: use Name for Plug and Slot name
This patch differentiates the Go and JSON interfaces to Plug and Slot
names. This comes after a discussion that wanted to eliminate the ugly
Plug.Plug "bork bork bork"-looking code without sacrificing the
unambiguity in the typeless JSON representation.

With this patch, the on-the-wire protocol remains the same, unambiguous
{"plug": "..."} and {"slot": "..."}.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>