many: change POST /2.0/interfaces to work with lists
This patch changes the REST API for interacting with interfaces to
operate on lists of plugs and slots. In practice nothing changes but
this patch opens the avenue for connecting or disconnecting multiple
plugs and sockets in one call.

The connect/disconnect/{add,remove}-{plug,slot} actions are constrained
to one plug/slot. This restriction can be lifted later by another patch,
without any changes to the wire format.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>