interfaces: support static and connections-specific security snippets
This patch changes the Interface interface to expose four
security-related methods. In addition to existing SlotSecuritySnippet()
and PlugSecuritySnippet() there are now StaticSlotSecuritySnippet() and
StaticPlugSecuritySnippet().

This change is driven by the realization that we can simplify security
and number of interfaces by disassociating some permissions from
established connections.

For example, a display server interface can now be just a single
interface rather than two. Consider this example:

1. There's a Mir snap, with the "mir" slot.
2. There's a xeyes snap with a "mir" plug.

The mir snap, simply because it has the mir *slot* gets to have access
to graphics cards and all the required machinery. Mir can start even
without xeyes running yet.

As a connection is made between xeyes and mir, a new set of permissions
are granted: Xeyes can now talk to mir socket.

The same example works with any managed, shared resource that the
managing snap needs to be able to control regardless of who is connected
at a particular moment.

This patch also changes test interface that is use for testing to
support the extra methods.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>