snappy: switch to per-application udev profile names
This patch fixes an older issue where udev "tag" or profile was global
to a snap. This mean that it was impossible to have different
permissions for character and block devices among apps in one snap. This
patch corrects this by switching to udev profile/tag following this
template:

    "${snap.name}.${snap.app}.${snap.origin}"

There are several test adjustments to make some related tests more
readable. In the past those tests would depend on doing blind string
replacement in expected outcome of another test that didn't work very
well after this change.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>