Move Series() to the core assertion types.
The `series` header field is required for all assertion types to support
the `account-key` lookup for the assertion's `authority-id`.

Therefore `series` is so core to the API it makes sense to:

- add Series() to the Assertion interface.
- implement Series() on the assertionBase type.
- check that `series` exists in Assemble().

For now, the check in Assemble() is commented out to avoid breaking all
the tests, allowing a more iterative approach.