snap,snappy,systemd: %s/\<SNAP_ORIGIN\>/SNAP_DEVELOPER/g
This patch begins the change away from "origin" towards the "snap
developer". The change starts with alterations of the public-facing
environment variables.

I will follow it with internal changes to rename "origin" to "developer"
and finally to remove the special-cased frameworks which currently,
artificially, always have an empty origin field. This change will be
intertwined with changes to application launchers that will gain the new
feature of having a special-cased short wrapper name of just Snap.Name
in the only case where App.Name is the same as Snap.Name. This will
allow things like docker not to be spelled out as docker.docker but
simply docker.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>