interfaces: add default seccomp template
This patch copies the default seccomp template that applies to all applications
in all snaps. Since we're keeping all the interface implementations in go it is
easier and arguably more correct to also hold the base policy in the same
place.

Subsequent patches will use this template and incorporate interface-specific
security snippets into it.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>