snappy: fix hw-assign to work with per-app udev tags
This patch changes 'snappy hw-assign' to work with the new
per-application udev tags.  The change is simple, instead of writing one
rule per snap, we write one rule for each app of the snap mentioned on
command line.

Unit tests are affected (there was some missing setup needed for one of
the tests, which now matters since we want to see snap meta-data) as
well as some part of the implementation (we sort the rules so that
behavior and tests are deterministic) but nothing major.

There should be no user-visible differences on the outside. This is only
meant to keep hw-assign working before developer mode and interfaces
make it entirely obsolete.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>