overlord/ifacestate: initial work on the interface manager.
This patch fills in some holes in the empty stubs that currently
constitute the interface manager. I've added the data structures that I
think will be required to perform all of the foreseen operations.

There's some simple interface repository initialization code that will
soon entirely replace the repository object held by the daemon (and fuel
all operations).

Lastly there are some TODO's I've added to Ensure() that show the
outline of what I think the manager is responsible for.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>