snappy: fix hw-assign to be compatible with hw-unassign
This patch tweaks my earlier change to hw-assign so that, from the
outside, it still looks like it operates on $snap.$origin (including all
the files created look this way). This is now compatible with the
unchanged hw-info and hw-unassign. Internally we still understand snap
name and snap origin separately. This change fixes three integration
test failures.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>