interfaces/janitor: add janitor module
This patch adds the janitor -- a flexible concept for maintaining a
subset of files in a directory. The janitor ensures that all the files
in the directory are exactly as they are expected to be and reports on
the changes made.

The janitor can be used to manage on-disk apparmor, seccomp, udev and
dbus configuration files. Reports about changes made can be used to
trigger udev configuration reload, load or unload apparmor profiles to
the kernel.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>