overlord/state: tweak locking in TaskRunner, add Settle()
The main aim of this patch is to add TaskRunner.Settle(), indented for
testing managers, which waits for all the task to finish without holding
the state lock so that they can actually do finish doing their work.

This is achieved by holding the local lock (specific to a particular
task runner) in both the existing Stop() and in the new Settle(). The
local lock now also protects modifications to the tomb map.

Since tombs are not protected by the state map, the goroutine that runs
the handler of a specific task can no longer clean up its own tomb. This
responsibility falls on Settle() and Stop(). To ensure that tombs don't
grow out of bound, Ensure() also cleans up tombs of dead goroutines
(which it can do because it already holds the local lock).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>