snap: fix app-{slot,port} binding algorithm
This patch fixes the binding algorithm to work properly.  Writing more
extensive tests made me realize the algorithm was broken.  We have to
explicitly track app-bound (so not global) plugs and slots.

For example, an application can have no slots but should not
automatically be bound to an app-bound slot used elsewhere.  Here, app2
should not be bound to slot foo. Slot foo is not global, it is only
bound to app1.

apps:
    app1:
        slots: [foo]
    app2:
slots:
    foo:
        ...

This patch also disambiguates variable names (so appName or slotName
rather than just name) so that following the code is easier.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>