snap: default interface to slot or plug name
This patch corrects the plug/slot analyzer to use the plug/slot name as
interface by default. When I read the spec I didn't see the comment
``Optional if <plug> == <interface>`` which, in context, means that this
YAML is indeed valid:

plugs:
    foo:
        attr: ...

The YAML snippet above defines a plug "foo" with interface of "foo".

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>