interfaces: make AddSnap panic if added snap clashes
This patch makes AddSnap mirror RemoveSnap in that they both assume to
work with a clean and ready environment. AddSnap is really expecting to
add a snap that doesn't exist yet and RemoveSnap (already) expecting to
remove a snap without any connections left. This makes the code above
those two functions easier to reason about (e.g. when can this fail).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>