overlord/snapstate: stop handling security tasks
This patch removes handlers for security tasks from the snap manager and
replaces them (but only during testing) with fake no-op handlers so that
task structure can be maintained.

This patch should be followed up with the complete removal of related
code (which is now dead). This affects both the snap manager backend and
large parts of the snappy/ package.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>