client,daemon,overlord: fix authentication
Due to a bug in overlord/auth, the daemon was blindly accepting
the authentication of anyone with an authentication header if
there was no auth state at all (nobody ever logged in).

This also includes a clean up of Logout, making it remove the
previously stored authentication details on success.