overlord/ifacestate: fix setup-profiles to use new snap revision for setup
This patch fixes a bug in setup-profiles where on snap upgrade, the
security would be still set up with the older revision of the snap.

This was caused by DisconnectSnap returning the list of snaps
(represented as snap.Info's) that are affected by the disconnect
operation. Those infos were subsequently used to setup the security of
each snap. This didn't work because snap.Info encodes the Revision which
changes on each upgrade.

Note that this bug is not a security vulnerability as apparmor uses the
revision in the actual profile so applications from both revisions are
simply unavailable.

Fixes: https://bugs.launchpad.net/snappy/+bug/1572463
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>