interfaces/builtin: specialize apparmor label based on bound apps
This patch specialized the bluez label used for dbus policy based on
how apps are bound to the bluez slot.

If there is exactly one app bound to the slot the label is precise, e.g.
"snap.bluez.bluez". If a subset of apps are bound to the slot (but not a
proper subset) then the label uses alternation, e.g.
"snap.bluez.{app1,app2}". Lastly if all the apps are bound to the bluez
slot then a wildcard is used, e.g. "snap.bluez.*".

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>