interfaces: add global gsettings interfaces (Closes #1576308)
When connected, applications can access the global gsettings of the user's
session. This interface is restricted because it gives privileged access to
sensitive information stored in gsettings (such as passwords) and allows
adjusting settings of other applications.

In the future gsettings will gain AppArmor integration to allow safe,
application-specific access to gsettings, but for now this interface can be
used to unblock people until that work is completed.