Make lsb-release handling compatible with non-Ubuntu OSes.
Use /etc/os-release, which is guaranteed to be part of base-files on both
Ubuntu and Debian, instead of /etc/lsb-release which doesn't exist at all
on Debian.

If no UBUNTU_CODENAME field is found, default to 'xenial', the codename for
the most recent Ubuntu LTS.  This field is used for downloads of Ubuntu
rootfses from LXD, so must be an Ubuntu codename, not the codename for
another distribution.