overlord/ifacestate: enable apparmor security backend on ubuntu
Ubuntu carries all the relevant apparmor support patches in the kernel
(as well as userspace tools) so it can use all of the apparmor security
fatures. Other distributions may not have this enabled so as a first
safe option they don't have apparmor and will (soon) require --devmode
installation to work.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>