snap: add parsing support for hooks.
Add support for an optional "hooks" property in the YAML that follows
the following format:

    hooks:
      hook-name
        plugs: [plug-name]

Note that the plugs are optional. Also note that hooks are factored into
the equation when considering whether or not a global plug is unbound--
if it's bound to either an app OR a hook it's considered bound for both,
in which case the plug will not be bound to all apps and hooks.

Updates: #1586465

Signed-off-by: Kyle Fazzari <kyle@canonical.com>