Use autotools build system
This patch discards the hand-crafted build system in favour of autotools.
This change makes it easier to build the source code and to begin adding
conditional features.

The seccomp.[ch] file was renamed to seccomp_utils.[ch] in order not to
clash with the system-wide seccomp.h. This is required because of how
autotools pass the -I. flag by default.

Unit tests were converted to autotool standards.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>