Have --disable-confinement disable *all* confinement
This patch changes the aforementioned configure switch to disable udev
device tagging, installation of udev tagging helper, use of seccomp, use
of private /tmp, use of private /dev/pts and /dev/ptmx as well as use of
apparmor.

With this option selected, the only thing that remains of out
snap-confine is the arrangement of the core snap bind mounts so that
applications can still run.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>