snap,cmd: add hook support to snap run.
The `snap run` command currently supports running the apps contained
within a snap. This commit adds the ability to run the hooks contained
within a snap as well, using the --hook option. By default it'll run
hooks on the currently-active revision of a snap, but it also supports
running hooks on a specific revision of a snap, using the --revision
option.

Updates: #1586465

Signed-off-by: Kyle Fazzari <kyle@canonical.com>