Reset PATH on --enable-rootfs-is-core-snap
This patch ensures that PATH has values sensible for the core snap. Whatver
the host PATH may contain, the core snap has predictable directories and snap
applications may rely on the PATH lookup to find utilities in /bin. This is
especially important on host systems that use a different root filesystem
arrangement (e.g. when /bin is a symlink to /usr/bin/ and thus doesn't
explicitly show up on PATH).

Signed-off-by: Zygmunt Krynicki <zkrynicki@gmail.com>