Improve detection of nvidia driver
This patch changes how nvidia driver is detected. Instead of looking at one
location that is specific to a given distribution, all such locations (from the
assorted set of distributions) are examined.

This will later be entirely replaced by snapd mount profiles but for now it is
a quick way to get to working nvidia support. In the end only the source of the
data will change.

Signed-off-by: Zygmunt Krynicki <zkrynicki@gmail.com>