Completely re-do nvidia driver support
This patch changes how nvidia driver is supported. The support differs
depending on the distribution via a compile-time configuration option.

On Ubuntu nothing is changed, we still look for /usr/lib/nvidia-... and
bind-mount it to /var/lib/snapd/lib/gl. This behavior is now enabled by
--enable-nvidia-ubuntu. It is off by default.

On Arch a totally different approach is used. A tmpfs is created in
/var/lib/snapd/lib/gl so that it can be populated with symlinks to
/var/lib/snapd/hostfs. The confinement program contains a list of
libraries that comprise the nvidia proprietary driver. Only those
libraries are exposed. This behavior is now enabled by
--enable-nvidia-arch. It is off by default.

The arch support was tested on various drivers, 304 and 340 drivers work
correctly out of the box. Sadly the most recent driver (364) does not
work yet and need more hacking.

This also depends on the new hostfs directory to exist in the core snap.
It was added to the build system and should be available as of the next
release of the core snap.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>