Exit when we cannot remain in the current working directory
This patch evolves the approach to current directory treatment. After
some discussion with the security team I'm now convinced that it should
always refuse to run if the current directory cannot be retained. It is
better to fail early and loudly in those few cases where this will
happen.

Note that this is separate from being able to read or write to the
current directory, this is simply related to the fact that the crafted
chroot that snap execute in cannot contain all the locations that are
present on the outside.

This patch changes the old behavior of relocating to the root directory
to an explicit failure message followed by exit from snap-confine. The
regression test was updated to match this.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>